+++ src/wagtail_lms/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
+from django.utils.module_loading import import_string
 
+from . import conf
 from .models import (
     CourseEnrollment,
     H5PActivity,
@@ -12,8 +14,10 @@ from .models import (
     SCORMPackage,
 )
 
+_DEFAULT_SCORM_ADMIN_CLASS_PATH = "wagtail_lms.admin.SCORMPackageAdmin"
+_DEFAULT_H5P_ADMIN_CLASS_PATH = "wagtail_lms.admin.H5PActivityAdmin"
 
-@admin.register(SCORMPackage)
+
 class SCORMPackageAdmin(admin.ModelAdmin):
     list_display = ("title", "version", "created_at", "launch_url")
     list_filter = ("version", "created_at")
@@ -27,14 +31,12 @@ class SCORMPackageAdmin(admin.ModelAdmin):
     )
 
 
-@admin.register(CourseEnrollment)
 class CourseEnrollmentAdmin(admin.ModelAdmin):
     list_display = ("user", "course", "enrolled_at", "completed_at")
     list_filter = ("enrolled_at", "completed_at")
     search_fields = ("user__username", "course__title")
 
 
-@admin.register(SCORMAttempt)
 class SCORMAttemptAdmin(admin.ModelAdmin):
     list_display = (
         "user",
@@ -48,7 +50,6 @@ class SCORMAttemptAdmin(admin.ModelAdmin):
     search_fields = ("user__username", "scorm_package__title")
 
 
-@admin.register(SCORMData)
 class SCORMDataAdmin(admin.ModelAdmin):
     list_display = ("attempt", "key", "value_preview", "timestamp")
     list_filter = ("key", "timestamp")
@@ -60,7 +61,6 @@ class SCORMDataAdmin(admin.ModelAdmin):
     value_preview.short_description = "Value Preview"
 
 
-@admin.register(H5PActivity)
 class H5PActivityAdmin(admin.ModelAdmin):
     list_display = ("title", "main_library", "created_at")
     list_filter = ("created_at",)
@@ -74,7 +74,6 @@ class H5PActivityAdmin(admin.ModelAdmin):
     )
 
 
-@admin.register(H5PAttempt)
 class H5PAttemptAdmin(admin.ModelAdmin):
     list_display = (
         "user",
@@ -88,22 +87,68 @@ class H5PAttemptAdmin(admin.ModelAdmin):
     search_fields = ("user__username", "activity__title")
 
 
-@admin.register(H5PXAPIStatement)
 class H5PXAPIStatementAdmin(admin.ModelAdmin):
     list_display = ("attempt", "verb_display", "timestamp")
     list_filter = ("timestamp",)
     search_fields = ("attempt__user__username", "verb", "verb_display")
 
 
-@admin.register(H5PContentUserData)
 class H5PContentUserDataAdmin(admin.ModelAdmin):
     list_display = ("attempt", "data_type", "sub_content_id", "updated_at")
     list_filter = ("data_type", "updated_at")
     search_fields = ("attempt__user__username", "attempt__activity__title", "data_type")
 
 
-@admin.register(LessonCompletion)
 class LessonCompletionAdmin(admin.ModelAdmin):
     list_display = ("user", "lesson", "completed_at")
     list_filter = ("completed_at",)
     search_fields = ("user__username", "lesson__title")
+
+
+def _import_admin_class(dotted_path, setting_name):
+    try:
+        admin_class = import_string(dotted_path)
+    except (ImportError, AttributeError) as exc:
+        raise ImportError(
+            f"Could not import '{dotted_path}' from {setting_name}"
+        ) from exc
+    if not issubclass(admin_class, admin.ModelAdmin):
+        raise TypeError(
+            f"{setting_name} must reference a Django ModelAdmin subclass; "
+            f"got '{dotted_path}'"
+        )
+    return admin_class
+
+
+def _register_django_admin():
+    if not conf.WAGTAIL_LMS_REGISTER_DJANGO_ADMIN:
+        return
+
+    if conf.WAGTAIL_LMS_SCORM_ADMIN_CLASS == _DEFAULT_SCORM_ADMIN_CLASS_PATH:
+        scorm_admin_class = SCORMPackageAdmin
+    else:
+        scorm_admin_class = _import_admin_class(
+            conf.WAGTAIL_LMS_SCORM_ADMIN_CLASS,
+            "WAGTAIL_LMS_SCORM_ADMIN_CLASS",
+        )
+
+    if conf.WAGTAIL_LMS_H5P_ADMIN_CLASS == _DEFAULT_H5P_ADMIN_CLASS_PATH:
+        h5p_admin_class = H5PActivityAdmin
+    else:
+        h5p_admin_class = _import_admin_class(
+            conf.WAGTAIL_LMS_H5P_ADMIN_CLASS,
+            "WAGTAIL_LMS_H5P_ADMIN_CLASS",
+        )
+
+    admin.site.register(SCORMPackage, scorm_admin_class)
+    admin.site.register(CourseEnrollment, CourseEnrollmentAdmin)
+    admin.site.register(SCORMAttempt, SCORMAttemptAdmin)
+    admin.site.register(SCORMData, SCORMDataAdmin)
+    admin.site.register(H5PActivity, h5p_admin_class)
+    admin.site.register(H5PAttempt, H5PAttemptAdmin)
+    admin.site.register(H5PXAPIStatement, H5PXAPIStatementAdmin)
+    admin.site.register(H5PContentUserData, H5PContentUserDataAdmin)
+    admin.site.register(LessonCompletion, LessonCompletionAdmin)
+
+
+_register_django_admin()